format fix

+++ contracts/mocks/IERC20old.sol
@@ -2,6 +2,10 @@
 pragma solidity 0.7.6;
 
 // some old ERC20 implementations do not return (bool) when transferFrom is called
-interface IERC20Old  {
-    function transferFrom(address from, address to, uint value) external;
+interface IERC20Old {
+    function transferFrom(
+        address from,
+        address to,
+        uint256 value
+    ) external;
 }
